edit validate controller

+++ app/Http/Controllers/AgendaController.php
@@ -30,6 +30,13 @@ class AgendaController extends Controller
      */
     public function store(Request $request)
     {
+        $request->validate([
+            'judul_agenda' => 'required|max:100',
+            'deskripsi_agenda' => 'required',
+            // 'gambar_agenda' => 'required',
+            'tanggal_agenda' => 'required',
+        ]);
+
         AgendaModel::create([
             'judul_agenda' => $request->judul_agenda,
             'deskripsi_agenda' => $request->deskripsi_agenda,
@@ -65,9 +72,14 @@ class AgendaController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $data = AgendaModel::find($id);
+        $request->validate([
+            'judul_agenda' => 'required|max:100',
+            'deskripsi_agenda' => 'required',
+            // 'gambar_agenda' => 'required',
+            'tanggal_agenda' => 'required',
+        ]);
 
-        $data->update([
+        AgendaModel::find($id)->update([
             'judul_agenda' => $request->judul_agenda,
             'deskripsi_agenda' => $request->deskripsi_agenda,
             'gambar_agenda' => '',
@@ -82,12 +94,17 @@ class AgendaController extends Controller
      */
     public function destroy(string $id)
     {
+        $check = AgendaModel::find($id);
+        if (!$check) {
+            return redirect('/RW/Agenda')->with('error', 'Data tidak ditemukan');
+        }
+
         try {
             AgendaModel::destroy($id);
 
             return redirect('/RW/Agenda')->with('success', 'Data berhasil dihapus');
-        } catch (e) {
-            return redirect('RW/Agenda')->with('error', 'Data gagal dihapus');
+        } catch (\Illuminate\Database\QueryException $e) {
+            return redirect('/RW/Agenda')->with('error', 'Data gagal dihapus');
         }
     }
 }
